Backend-FastAPI-TDD/pyproject.toml
[project]
name = "projeto-tdd"
version = "0.1.0"
description = ""
authors = [
    {name = "Pedro",email = "pedro.felipe11@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "pymongo (>=4.14.1,<5.0.0)"
]

[tool.poetry]
packages = [{include = "projeto_tdd", from = "src"}]


[tool.poetry.group.dev.dependencies]
ruff = "^0.12.11"
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
xfail_strict = true
junit_family = "xunit2"
addopts = [
    "-p", "no:warnings",
    "--strict-config",
    "--strict-markers",
    "--ignore=docs_src",
]


